separate cream and compound classifiers
